Expand and document API
